% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBP.RTM.R
\name{DBP.RTM}
\alias{DBP.RTM}
\title{Regression to the mean}
\usage{
DBP.RTM(
  mu = 90,
  sigma = 8,
  r = 0.76,
  n = 1000,
  limit = 95,
  TrueChange = 0,
  show.plot = TRUE,
  show.out = FALSE
)
}
\arguments{
\item{mu}{The mean DBP at baseline}

\item{sigma}{The standard deviation at baseline and at follow-up}

\item{r}{The pre-post correlation}

\item{n}{sample size}

\item{limit}{Limit for hypertension}

\item{TrueChange}{True Change from Baseline. Default is zero.}

\item{show.plot}{Add plot, default=TRUE}

\item{show.out}{print results, default=FALSE}
}
\value{
A list with
\itemize{
\item plot showing equality line and the linear regression slope for all observations
\item plot showing extreme subjects only with regression to the mean
\item A list with
\itemize{
\item summary I - Summary statistics for all observations
\item summary II - Summary statistics for extreme cases (with hypertension) only
\item correlation Pre-Post correlation or intraclass correlation
\item t-test all
\item t-test extreme
\item model Coefs from simple linear regression (that is, by "fixing" baseline), equals the correlation coefficient for standardized data
\item fraction of regression to the mean, RTM, equals (1-rho)
\item dataI all observations
\item dataII extreme cases
}}
}
\description{
Regression to the mean
}
\examples{
##Assume a true change of -10 (on diastolic blood presssure)
DBP.RTM(TrueChange=-10)
}
\references{
Senn, S. (2009) Three things that every medical writer should know about statistics. The Write Stuff, 18 (3). pp. 159-162. ISSN 1854-8466
}
\author{
meichtry
}
