% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTMexample.R
\docType{package}
\name{RegToMeanExample}
\alias{RegToMeanExample}
\title{RegToMeanExample: A package for visualising regression to the mean}
\description{
The RegToMeanExample package provides DBP.RTM:
}
\section{RegToMeanExmaple functions}{

DBP.RTM
}

